Consolidate state version logic, add helpers to simplify updates
