+++ cannon/multicannon/exec.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"slices"
 
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm/versions"
 )
@@ -21,7 +20,7 @@ var vmFS embed.FS
 const baseDir = "embeds"
 
 func ExecuteCannon(ctx context.Context, args []string, ver versions.StateVersion) error {
-	if !slices.Contains(versions.StateVersionTypes, ver) {
+	if !versions.IsValidStateVersion(ver) {
 		return errors.New("unsupported version")
 	}
